% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_m.R
\name{get_m}
\alias{get_m}
\title{Title}
\usage{
get_m(
  mvals,
  data,
  weight,
  Y_index,
  X_index,
  C_index,
  N,
  method = "MCEM",
  ini_alpha = NULL
)
}
\arguments{
\item{ini_alpha}{}
}
\value{

}
\description{
Title
}
