% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAapproxL.R
\name{LAapproxL}
\alias{LAapproxL}
\title{Choose smoothing parameter with marginal likelihood estimated by Laplace approximation}
\usage{
LAapproxL(m, XM, y, ini_alpha, confounding_factor, weight_col)
}
\arguments{
\item{weight_col}{}
}
\value{

}
\description{
Choose smoothing parameter with marginal likelihood estimated by Laplace approximation
}
