% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logFmatched.R
\name{logFmatched}
\alias{logFmatched}
\title{log-F(m,m)-penalized conditional likelihood inference by data augmentation}
\usage{
logFmatched(form, dat, m = 1)
}
\arguments{
\item{m}{}
}
\value{

}
\description{
log-F(m,m)-penalized conditional likelihood inference by data augmentation
}
